Added dashboardState
Added dashboardState to index, set function and state are passed down to dashboard. NOTE: React doesnt like the fact that the function is being passed down several layers, hence why in workspace an anonymous function is passed down instead. This is a temporary workaround and should probably be fixed later.